Fix bug of disappearing wall
The wall sometimes disappears when playerX or playerY is close to
integer. The variable 'somethingDone' was used to handle the situation
where there seems to be no colliding walls. However, both weird behavior
is cause by tiny errors of floating point. By adding a tiny number DELTA
to the comparison, this bug can be solved and the variable
'somethingDone' is no longer needed.